# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mare Nostrum API",
        "description": "Smart Mobility Flow — Batam-Singapore cross-border logistics tracking API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/ports": {
            "get": {
                "tags": [
                    "Ports"
                ],
                "summary": "Get paginated list of ports",
                "operationId": "8628b3fb47954a3226512c8a6184243d",
                "parameters": [
                    {
                        "name": "country",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "indonesia",
                                "singapore"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of ports retrieved successfully."
                    }
                }
            },
            "post": {
                "tags": [
                    "Ports"
                ],
                "summary": "Create a new port",
                "operationId": "b23afbab3baced76c2d35e62880d4358",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "country",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Batam Centre Ferry Terminal"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "indonesia",
                                        "enum": [
                                            "indonesia",
                                            "singapore"
                                        ]
                                    },
                                    "unlocode": {
                                        "type": "string",
                                        "example": "IDBTH"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.1312345
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 104.0532145
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Port created successfully."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/ports/{id}": {
            "get": {
                "tags": [
                    "Ports"
                ],
                "summary": "Get single port details",
                "operationId": "77d4e2504b1df2526105d00f2c55febf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Port ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Port details retrieved successfully."
                    },
                    "404": {
                        "description": "Port not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Ports"
                ],
                "summary": "Update existing port details",
                "operationId": "5502a37aa519e69332b282320c295b23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Port ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Batam Centre Ferry Terminal Updated"
                                    },
                                    "country": {
                                        "type": "string",
                                        "enum": [
                                            "indonesia",
                                            "singapore"
                                        ]
                                    },
                                    "unlocode": {
                                        "type": "string",
                                        "example": "IDBTH"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Port updated successfully."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ports"
                ],
                "summary": "Delete a port",
                "operationId": "66d16a7e557e3ef0564b34dfe85af065",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Port ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Port deleted successfully."
                    },
                    "422": {
                        "description": "Cannot delete port referenced in trips."
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Send as: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}